requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
DAWG-Python==0.7.2
docopt==0.6.2
greenlet==2.0.2
idna==3.4
psycopg2==2.9.5
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-dotenv==1.0.0
requests==2.28.2
SQLAlchemy==2.0.7
typing_extensions==4.5.0
urllib3==1.26.15
vk-api==11.9.9